chore: introduce capabilities argument (#135)
