+++ src/tools/tool.ts
@@ -18,6 +18,8 @@ import type { ImageContent, TextContent } from '@modelcontextprotocol/sdk/types'
 import type { JsonSchema7Type } from 'zod-to-json-schema';
 import type { Context } from '../context';
 
+export type ToolCapability = 'core' | 'tabs' | 'pdf' | 'history' | 'wait' | 'files' | 'install';
+
 export type ToolSchema = {
   name: string;
   description: string;
@@ -30,6 +32,7 @@ export type ToolResult = {
 };
 
 export type Tool = {
+  capability: ToolCapability;
   schema: ToolSchema;
   handle: (context: Context, params?: Record<string, any>) => Promise<ToolResult>;
 };
